+++ llm_guard_api/app/app.py
@@ -2,28 +2,34 @@ import argparse
 import asyncio
 import concurrent.futures
 import time
+from typing import Annotated
 
 import structlog
-from cache import InMemoryCache
-from config import get_config
-from fastapi import FastAPI, HTTPException, status
+from fastapi import Depends, FastAPI, HTTPException, Response, status
 from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import JSONResponse
-from schemas import (
-    AnalyzeOutputRequest,
-    AnalyzeOutputResponse,
-    AnalyzePromptRequest,
-    AnalyzePromptResponse,
-)
+from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
+from slowapi import Limiter, _rate_limit_exceeded_handler
+from slowapi.errors import RateLimitExceeded
+from slowapi.middleware import SlowAPIMiddleware
+from slowapi.util import get_remote_address
 from starlette.exceptions import HTTPException as StarletteHTTPException
 from util import configure_logger
 
 from llm_guard import scan_output, scan_prompt
 from llm_guard.vault import Vault
 
-version = "0.0.6"
+from .cache import InMemoryCache
+from .config import get_config
+from .schemas import (
+    AnalyzeOutputRequest,
+    AnalyzeOutputResponse,
+    AnalyzePromptRequest,
+    AnalyzePromptResponse,
+)
+from .version import __version__
 
 vault = Vault()
 
@@ -35,36 +41,36 @@ scanners_config_file = args.config
 config = get_config(vault, scanners_config_file)
 
 LOGGER = structlog.getLogger(__name__)
-log_level = config["app"]["log_level"]
+log_level = config.app.get("log_level")
 is_debug = log_level == "DEBUG"
 configure_logger(log_level)
 
 
 def create_app():
     cache = InMemoryCache(
-        max_size=int(config["app"]["cache_ttl"]),
-        expiration_time=int(config["app"]["cache_ttl"]),
+        max_size=int(config.cache.get("max_size")),
+        expiration_time=int(config.cache.get("ttl")),
     )
 
-    if config["app"]["scan_fail_fast"]:
+    if config.app.get("scan_fail_fast"):
         LOGGER.debug("Scan fail_fast mode is enabled")
 
     app = FastAPI(
-        title=config["app"]["name"],
+        title=config.app.get("name"),
         description="API to run LLM Guard scanners.",
         debug=is_debug,
-        version=version,
+        version=__version__,
         openapi_url="/openapi.json" if is_debug else None,  # hide docs in production
     )
 
-    register_middlewares(app)
-
-    register_routes(app, cache, config["input_scanners"], config["output_scanners"])
+    register_routes(app, cache, config.input_scanners_loaded, config.output_scanners_loaded)
 
     return app
 
 
-def register_middlewares(app: FastAPI):
+def register_routes(
+    app: FastAPI, cache: InMemoryCache, input_scanners: list, output_scanners: list
+):
     app.add_middleware(
         CORSMiddleware,
         allow_origins=["*"],
@@ -73,25 +79,46 @@ def register_middlewares(app: FastAPI):
         allow_headers=["Authorization", "Content-Type"],
     )
 
+    limiter = Limiter(key_func=get_remote_address, default_limits=[config.rate_limit.get("limit")])
+    app.state.limiter = limiter
+    app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)
+    if bool(config.rate_limit.get("enabled")):
+        app.add_middleware(SlowAPIMiddleware)
 
-def register_routes(
-    app: FastAPI, cache: InMemoryCache, input_scanners: list, output_scanners: list
-):
+    security = HTTPBearer()
+
+    async def is_token_valid(
+        credentials: Annotated[HTTPAuthorizationCredentials, Depends(security)]
+    ) -> bool:
+        if credentials.credentials != config.auth.get("token"):
+            raise HTTPException(status_code=401, detail="Invalid API key")
+
+        return True
+
     @app.get("/", tags=["Main"])
+    @limiter.exempt
     async def read_root():
         return {"name": "LLM Guard API"}
 
     @app.get("/healthz", tags=["Health"])
+    @limiter.exempt
     async def healthcheck():
         return JSONResponse({"status": "alive"})
 
     @app.get("/readyz", tags=["Health"])
+    @limiter.exempt
     async def liveliness():
         return JSONResponse({"status": "ready"})
 
-    @app.post("/analyze/output", tags=["Analyze"])
+    @app.post(
+        "/analyze/output",
+        tags=["Analyze"],
+        response_model=AnalyzeOutputResponse,
+        status_code=status.HTTP_200_OK,
+    )
+    @limiter.limit("5/minute")
     async def analyze_output(
-        request: AnalyzeOutputRequest,
+        request: AnalyzeOutputRequest, _: Annotated[bool, Depends(is_token_valid)]
     ) -> AnalyzeOutputResponse:
         LOGGER.debug("Received analyze output request", request=request)
 
@@ -106,9 +133,9 @@ def register_routes(
                         output_scanners,
                         request.prompt,
                         request.output,
-                        config["app"]["scan_fail_fast"],
+                        bool(config.app.get("scan_fail_fast")),
                     ),
-                    timeout=float(config["app"]["scan_output_timeout"]),
+                    timeout=float(config.app.get("scan_output_timeout")),
                 )
 
                 response = AnalyzeOutputResponse(
@@ -129,18 +156,28 @@ def register_routes(
 
         return response
 
-    @app.post("/analyze/prompt", tags=["Analyze"])
+    @app.post(
+        "/analyze/prompt",
+        tags=["Analyze"],
+        response_model=AnalyzePromptResponse,
+        status_code=status.HTTP_200_OK,
+    )
     async def analyze_prompt(
         request: AnalyzePromptRequest,
+        _: Annotated[bool, Depends(is_token_valid)],
+        response: Response,
     ) -> AnalyzePromptResponse:
         LOGGER.debug("Received analyze prompt request", request=request)
-        cached_result = cache.get(request.prompt)
 
+        cached_result = cache.get(request.prompt)
         if cached_result:
             LOGGER.debug("Response was found in cache")
-            cached_result["is_cached"] = True
+
+            response.headers["X-Cache-Hit"] = "true"
 
             return AnalyzePromptResponse(**cached_result)
+
+        response.headers["X-Cache-Hit"] = "false"
 
         with concurrent.futures.ThreadPoolExecutor() as executor:
             loop = asyncio.get_event_loop()
@@ -152,9 +189,9 @@ def register_routes(
                         scan_prompt,
                         input_scanners,
                         request.prompt,
-                        config["app"]["scan_fail_fast"],
+                        bool(config.app.get("scan_fail_fast")),
                     ),
-                    timeout=float(config["app"]["scan_prompt_timeout"]),
+                    timeout=float(config.app.get("scan_prompt_timeout")),
                 )
 
                 response = AnalyzePromptResponse(
@@ -178,7 +215,7 @@ def register_routes(
         return response
 
     @app.on_event("shutdown")
-    def shutdown_event():
+    async def shutdown_event():
         LOGGER.info("Shutting down app...")
 
     @app.exception_handler(StarletteHTTPException)
@@ -208,7 +245,7 @@ def run_app():
     uvicorn.run(
         app,
         host="0.0.0.0",
-        port=int(config["app"]["port"]),
+        port=int(config.app.get("port")),
         server_header=False,
         log_level=log_level.lower(),
         proxy_headers=True,